// src/game/enums/CardType.ts
/**
 * Tipos de cartas do jogo Cunoku
 */
export enum CardType {
  /** Rei - valor 0 */
  REI = 'REI',
  /** Coringa - valor -1 */
  CORINGA = 'CORINGA',
  /** Ás - valor 1 */
  AS = 'AS',
  /** Valete - valor 11 */
  VALETE = 'VALETE',
  /** Rainha - valor 12 */
  RAINHA = 'RAINHA',
  /** Cartas numéricas (5-12) */
  NUMERO = 'NUMERO',
}